fin du cours

+++ backend/controllers/sauce.js
@@ -1,9 +1,13 @@
 const sauceSchema = require("../database/models/sauce");
 
 exports.createSauce = (req, res, next) => {
-  delete req.body._id;
+  const sauceObject = JSON.parse(req.body.sauce);
+  delete sauceObject._id;
   const sauce = new sauceSchema({
-    ...req.body,
+    ...sauceObject,
+    imageUrl: `${req.protocol}://${req.get("host")}/images/${
+      req.file.filename
+    }`,
   });
   sauce
     .save()
@@ -11,17 +15,38 @@ exports.createSauce = (req, res, next) => {
 };
 
 exports.modifySauce = (req, res, next) => {
+  const sauceObject = req.file
+    ? {
+        ...JSON.parse(req.body.sauce),
+        imageUrl: `${req.protocol}://${req.get("host")}/images/${
+          req.file.filename
+        }`,
+      }
+    : { ...req.body };
   sauceSchema
-    .updateOne({ _id: req.params.id }, { ...req.body, _id: require.params.id })
+    .updateOne(
+      { _id: req.params.id },
+      { ...sauceObject, _id: require.params.id }
+    )
     .then(() => res.status(200).json({ message: "Sauce modifiée !" }))
     .catch(() => res.status(400).json({ error }));
 };
 
 exports.deleteSauce = (req, res, next) => {
-  sauceSchema
-    .deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: "Sauce supprimée !" }))
-    .catch((error) => res.status(400).json({ error }));
+  sauceSchema.findOne({ _id: req.params.id }).then((sauce) => {
+    if (!sauce) {
+      return res.status(404).json({
+        error: new Error("Objet non trouvé !"),
+      });
+    }
+    if (sauce.userId !== req.auth.userId) {
+      res.status(401).json({ error: new Error("Requête non authorisée !") });
+    }
+    sauceSchema
+      .deleteOne({ _id: req.params.id })
+      .then(() => res.status(200).json({ message: "Sauce supprimée !" }))
+      .catch((error) => res.status(400).json({ error }));
+  });
 };
 
 exports.getOneSauce = (req, res, next) => {
